% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addremove_dot.R
\name{remove_dot}
\alias{remove_dot}
\title{Remove dot from an ICD-10 code}
\usage{
remove_dot(x)
}
\arguments{
\item{x}{A valid ICD-10 code with a dot}
}
\value{
A valid ICD-10 code without a dot included
}
\description{
\code{remove_dot()} removes a dot from the ICD-10 code
}
\examples{
remove_dot("F32.0")

remove_dot("F32")

icd10cm \%>\% mutate(remove_dot(add_dot(icd10cm_code)))
}
