% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_icd10cm.R
\name{sample_icd10cm}
\alias{sample_icd10cm}
\title{Sample ICD-10-CM diagnosis codes}
\usage{
sample_icd10cm(n = 100, p = 10, start_code = "A00", end_code = "Z99")
}
\arguments{
\item{n}{number of ICD-10-CM to sample}

\item{p}{number of distinct patients to sample}

\item{start_code}{Lowest ICD-10-CM code to sample (default: A00)}

\item{end_code}{Highest ICD-10-CM code to sample (default: Z99)}
}
\value{
tibble
}
\description{
#' @description
\code{sample_icd10cm} returns a tibble of a random sample of ICD-10-CM diagnosis.
Currently there is uniform sampling across all billable ICD-10 codes.  Dates
and patient IDs are also randomly generated
}
\examples{
sample_icd10cm(n=1000)
}
