% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes_between.R
\name{codes_between}
\alias{codes_between}
\title{List all codes between two sets of codes}
\usage{
codes_between(start_code, end_code, expand_end_code = TRUE)
}
\arguments{
\item{start_code}{A string with the starting ICD-10-CM code}

\item{end_code}{A string with the ending ICD-10-CM code}

\item{expand_end_code}{Boolean indicating whether to include all
children of the end code}
}
\value{
Tibble of all matching ICD-10-CM rows from icd10cm
}
\description{
\code{codes_between} returns a tibble of all ICD-10-CM codes occuring
between two codes.  The function uses the order number included
in the source data from the CDC.
}
\examples{
codes_between("F32", "F33")
}
