% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_code.R
\name{expand_code}
\alias{expand_code}
\title{List all codes between two sets of codes}
\usage{
expand_code(x)
}
\arguments{
\item{x}{a valid ICD-10-CM code}
}
\value{
Tibble of all matching ICD-10-CM rows from icd10cm
}
\description{
\code{expand_code} returns a tibble of all ICD-10-CM codes that
are child codes of the provided code.
}
\examples{
expand_code("F32")
}
