% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10cm}
\alias{icd10cm}
\title{ICD-10-CM Codes.}
\format{
A data frame with seven variables:
\describe{
\item{\code{order_number}}{file order from source data}
\item{\code{icd10cm_code}}{ICD-10-CM code without a dot}
\item{\code{valid_billing_code}}{Is the code a valid billing code 1=Yes, 0=No}
\item{\code{icd10cm_short_description}}{ICD-10-CM short description}
\item{\code{icd10cm_long_description}}{ICD-10-CM long description}
\item{\code{poa_exempt}}{Flag to indicate if code is exempt from present on admission (POA) rules}
\item{\code{chronic indicator}}{Flag to indicate if code is acute or chronic based on HCUP Chronic Care Indicator data}
}
}
\source{
\url{ftp://ftp.cdc.gov/pub/Health_Statistics/NCHS/Publications/ICD10CM/2022/Code Descriptions zip.zip}

\url{ftp://ftp.cdc.gov/pub/Health_Statistics/NCHS/Publications/ICD10CM/2022/POAexemptCodesFY22.zip}

\url{https://www.hcup-us.ahrq.gov/toolssoftware/chronic_icd10/CCI-ICD10CM-v2021-1.zip}
}
\usage{
icd10cm
}
\description{
ICD-10-CM diagnosis codes
}
\keyword{datasets}
