% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_icd10cm.R
\docType{data}
\name{icd10cm_chapters}
\alias{icd10cm_chapters}
\title{ICD-10-CM Chapters}
\format{
A data frame with 5 variables:
\describe{
\item{\code{chapter_num}}{number of chapter}
\item{\code{chapter_abbr}}{chapter abbreviation}
\item{\code{chapter_desc}}{chapter long description}
\item{\code{code_start}}{first code in the chapter}
\item{\code{code_end}}{last code in the chapter}
}
}
\usage{
icd10cm_chapters
}
\description{
ICD-10-cM diagnosis code chapters (body systems)
}
\keyword{datasets}
