% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-icd10pcs.R
\docType{data}
\name{icd10pcs}
\alias{icd10pcs}
\title{ICD-10-PCS Codes.}
\format{
A data frame with 6 variables:
\describe{
\item{\code{order_number}}{file order from source data}
\item{\code{icd10pcs_code}}{ICD-10-PCS code}
\item{\code{valid_billing_code}}{Is the code a valid billing
code 1=Yes, 0=No}
\item{\code{icd10pcs_short_description}}{ICD-10-PCS short description}
\item{\code{icd10pcs_long_description}}{ICD-10-PCS long description}
\item{\code{procedure_class}}{class of procedure based on HCUP Procedure
Class data}
}
}
\source{
\url{https://www.cms.gov/files/zip/2022-icd-10-pcs-order-file-long-
and-abbreviated-titles.zip}

\url{https://www.hcup-us.ahrq.gov/toolssoftware/procedureicd10
/ProcedureClasses_v2021-2.zip}
}
\usage{
icd10pcs
}
\description{
ICD-10-PCS procedure codes
}
\details{
Procedure class is derived from the HCUP Procedure Class file.  From the
user guide, codes are divided into 4 classes:\tabular{ll}{
   \strong{procedure_class} \tab \strong{description} \cr
   Minor Diagnostic \tab Nonoperating room procedures that are diagnostic (e.g., B244ZZZ, Ultrasonography of Right Heart) \cr
   Minor Therapeutic \tab Nonoperating room procedures that are therapeutic (e.g., 02HQ33Z, Insertion of Infusion Device into Right Pulmonary Artery, Percutaneous Approach) \cr
   Major Diagnostic \tab Procedures that are considered operating room procedures that are performed for diagnostic reasons (e.g., 02BV0ZX, Excision of Superior Vena Cava, Open Approach, Diagnostic) \cr
   Major Therapeutic \tab Procedures that are considered operating room procedures that are performed for therapeutic reasons (e.g., 0210093, Bypass Coronary Artery, One Site from Coronary Artery with Autologous Venous Tissue, Open Approach). \cr
}
}
\keyword{datasets}
