% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-icd10pcs.R
\docType{data}
\name{icd10pcs_sections}
\alias{icd10pcs_sections}
\title{ICD-10-PCS Code Sections.}
\format{
A data frame with 2 variables:
\describe{
\item{\code{section_digit}}{first digit from ICD-10-PCS codes}
\item{\code{section_description}}{ICD-10-PCS section description}
}
}
\usage{
icd10pcs_sections
}
\description{
ICD-10-PCS procedure code top-level sections
}
\keyword{datasets}
